refactor: created share move regexp.ts in it; rename dbcore to cores
