--- src/database/dbcore/index.ts
@@ -1,3 +1,0 @@
-export interface Database {
-    id: number;
-}
